Refactor : skills 타입 파일도 이름 변경
