+++ src/components/about/About.tsx
@@ -2,7 +2,7 @@ import Skills from '@/components/about/Skills'
 import Education from '@/components/about/Education'
 import React from 'react'
 import { EducationItemData } from '@/types/education'
-import { SkillItemData } from '@/types/skill'
+import { SkillItemData } from '@/types/skills'
 import { getEducationList, getSkillList } from '@/lib/api/aboutApi'
 
 export default async function About() {
